+++ src/Vector2D.cpp
@@ -1,7 +1,7 @@
 #include "Vector2D.h"
 
 Vector2D::Vector2D():x(0),y(0){}
-Vector2D::Vector2D(float xpos,float ypos):x(xpos),y(ypos){}
+Vector2D::Vector2D(int xpos,int ypos):x(xpos),y(ypos){}
 Vector2D::~Vector2D(){}
 
 Vector2D Vector2D :: operator+(const Vector2D& op)const{
@@ -23,8 +23,8 @@ Vector2D& Vector2D ::operator+=(const Vector2D& op){
             return *this;
         }
 
-float& Vector2D :: getx() {return this->x;}
-float& Vector2D :: gety() {return this->y;}
+int& Vector2D :: getx() {return this->x;}
+int& Vector2D :: gety() {return this->y;}
 
-void Vector2D :: setx(float xpos) {this->x = xpos;}
-void Vector2D :: sety(float ypos) {this->y = ypos;}
+void Vector2D :: setx(int xpos) {this->x = xpos;}
+void Vector2D :: sety(int ypos) {this->y = ypos;}
